+++ disassembler.cpp
@@ -130,9 +130,9 @@ powerpc_decompose(const uint8_t *data, int size, uint32_t addr, bool lil_end,
 	res->handle = handle;
 
 	/* call */
-	int n = cs_disasm(handle, data, size, addr, 1, &insn);
+	size_t n = cs_disasm(handle, data, size, addr, 1, &insn);
 	if(n != 1) {
-		MYLOG("ERROR: cs_disasm() returned %d (cs_errno:%d)\n", n, cs_errno(handle));
+		MYLOG("ERROR: cs_disasm() returned %" PRIdPTR " (cs_errno:%d)\n", n, cs_errno(handle));
 		goto cleanup;
 	}
